+++ packages/orq-rc/src/models/components/index.ts
@@ -29,6 +29,7 @@ export * from "./budgetperiod.js";
 export * from "./budgetrejection.js";
 export * from "./budgetscope.js";
 export * from "./budgetscopekind.js";
+export * from "./budgetsortfield.js";
 export * from "./checkbudgetsrequest.js";
 export * from "./checkbudgetsresponse.js";
 export * from "./conversationparam.js";