+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-17T07:33:46.419Z",
+    "2026-06-17T09:58:08.470Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {