+++ packages/orq-rc/RELEASES.md
@@ -10059,3 +10059,13 @@ Based on:
 - [typescript v4.11.0-rc.27] packages/orq-rc
 ### Releases
 - [NPM v4.11.0-rc.27] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.27 - packages/orq-rc
+
+## 2026-06-17 10:01:01
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.781.0 (2.907.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.11.0-rc.28] packages/orq-rc
+### Releases
+- [NPM v4.11.0-rc.28] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.28 - packages/orq-rc